fix(server): correct nil check for diagnostics
diags might not be nil but instead merely empty as it is a slice.
the check should be minimally `len(diags) > 0` not `nil`.

even if it is not empty, it may only contain non-error diagnostics.

```golang
// This helper aims to mimic the go error practices of if err != nil. After any
// operation that returns Diagnostics, check that it HasError and bubble up the
// stack.
```

unexported helpers within may hold the `nil` invariant, but terraform
plugin helpers etc. may not.